refactor(sandbox): split discovered-definition + execute + runtime
